+++ packages/shortest/src/ai/tools/index.ts
@@ -103,6 +103,10 @@ export const AITools = [
     },
   },
   {
+    type: "bash_20241022",
+    name: "bash",
+  },
+  {
     name: "upload_file",
     description: "Set the upload file to the file specified",
     input_schema: {